# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ticket Platform API
  version: 0.1.0
  description: |
    Initial API contract for gateway public endpoints and internal service endpoints.
servers:
  - url: https://api.example.com
    description: Public API Gateway
  - url: http://localhost:3000
    description: Local API Gateway
...
paths:
...
  /v1/marketplace/listings:
    get:
      tags: [Marketplace]
      summary: Browse resale listings
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Listing collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListingListResponse"
    post:
      tags: [Marketplace]
      summary: Create resale listing (requires KYC)
      parameters:
        - $ref: "#/components/parameters/XUserId"
        - $ref: "#/components/parameters/XKycStatus"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceCreateListingInput"
      responses:
        "200":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListingResponse"
        "400":
          description: Invalid payload
        "401":
          description: Missing user context
        "403":
          description: KYC not approved
        "409":
          description: Token already listed
...
  /v1/marketplace/me/sales:
    get:
      tags: [Marketplace]
      summary: List seller completed sales
      parameters:
        - $ref: "#/components/parameters/XUserId"
      responses:
        "200":
          description: Completed sales list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceSalesResponse"
        "401":
          description: Missing user context
  /v1/internal/marketplace/settlements/finalize:
    post:
      tags: [Marketplace]
      summary: Internal settlement finalization and submit trigger
      parameters:
        - $ref: "#/components/parameters/InternalApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceSettlementFinalizeInput"
      responses:
        "200":
          description: Settlement submitted or returned from idempotent ledger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceSettlementFinalizeResponse"
        "400":
          description: Invalid settlement payload
        "401":
          description: Invalid internal API key
...
components:
  parameters:
    XUserId:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
      description: Simulated user identity header used by current skeleton.
    XKycStatus:
      name: x-kyc-status
      in: header
      required: false
      schema:
        type: string
      description: Simulated KYC status header. Listing creation requires `approved`.
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: false
      schema:
        type: string
      description: Optional idempotency key for mutation endpoints.
...
    InternalApiKey:
      name: x-internal-api-key
      in: header
      required: true
      schema:
        type: string
      description: Internal service-to-service API key.
  schemas:
...
    MarketplaceCreateListingInput:
      type: object
      properties:
        tokenId:
          type: string
        eventId:
          type: string
        originalPrice:
          type: number
        askPrice:
          type: number
        sellerWalletAddress:
          type: string
      required: [tokenId, eventId, originalPrice, askPrice, sellerWalletAddress]
    MarketplaceListing:
      type: object
      properties:
        id:
          type: string
        tokenId:
          type: string
        eventId:
          type: string
        sellerUserId:
          type: string
        sellerWalletAddress:
          type: string
        originalPrice:
          type: number
        askPrice:
          type: number
        currency:
          type: string
          enum: [VND]
        status:
          type: string
          enum: [active, cancelled, completed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        buyerUserId:
          type: string
        paymentId:
          type: string
        settlementId:
          type: string
    MarketplaceListingResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MarketplaceListing"
    MarketplaceListingListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/MarketplaceListing"
...
    MarketplaceSalesItem:
      type: object
      properties:
        listingId:
          type: string
        paymentId:
          type: string
        settlementId:
          type: string
        escrowDataHash:
          type: string
        completeSaleRequestId:
          type: string
        completedAt:
          type: string
          format: date-time
        tokenId:
          type: string
        eventId:
          type: string
        buyerUserId:
          type: string
        askPrice:
          type: number
        currency:
          type: string
    MarketplaceSalesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/MarketplaceSalesItem"
    MarketplaceSettlementFinalizeInput:
      type: object
      properties:
        version:
          type: integer
          enum: [1]
        settlementId:
          type: string
        listingId:
          type: string
        paymentId:
          type: string
        tokenId:
          type: string
        seller:
          type: string
        buyer:
          type: string
        grossAmount:
          type: number
        sellerAmount:
          type: number
        platformFee:
          type: number
        organizerRoyalty:
          type: number
        currency:
          type: string
        gateway:
          type: integer
        gatewayReference:
          type: string
        settledAt:
          type: integer
        nonce:
          type: string
      required:
        [
          version,
          settlementId,
          listingId,
          paymentId,
          tokenId,
          seller,
          buyer,
          grossAmount,
          sellerAmount,
          platformFee,
          organizerRoyalty,
          currency,
          gateway,
          gatewayReference,
          settledAt,
          nonce
        ]
    MarketplaceSettlementFinalizeRecord:
      type: object
      properties:
        settlementId:
          type: string
        listingId:
          type: string
        paymentId:
          type: string
        escrowDataHash:
          type: string
        submitTxHash:
          type: string
        status:
          type: string
          enum: [submitted]
        submittedAt:
          type: string
          format: date-time
    MarketplaceSettlementFinalizeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MarketplaceSettlementFinalizeRecord"
...